+++ library/src/main/jni/com_twilio_video_ConnectOptions.h
@@ -36,6 +36,7 @@ JNIEXPORT jlong JNICALL Java_com_twilio_video_ConnectOptions_nativeCreate
          jobjectArray,
          jobject,
          jboolean,
+         jboolean,
          jlong,
          jobjectArray,
          jobjectArray,